Moved mod picker out into a new common loadout folder as well as its dependencies.
