+++ src/app/loadout-drawer/SavedModCategory.tsx
@@ -1,6 +1,6 @@
 import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
-import LockedModIcon from 'app/loadout-builder/filter/LockedModIcon';
+import LockedModIcon from 'app/loadout/loadout-ui/LockedModIcon';
 import React from 'react';
 import { AddButton } from './Buttons';
 import styles from './SavedModCategory.m.scss';
