+++ src/app/loadout/loadout-ui/ClosableContainer.m.scss
