+++ src/app/loadout/loadout-ui/ClosableContainer.tsx
