+++ src/app/loadout/loadout-ui/LockedModIcon.m.scss.d.ts
