+++ src/app/loadout/loadout-ui/LockedModIcon.tsx
@@ -2,7 +2,7 @@ import { D2ManifestDefinitions } from 'app/destiny2/d2-definitions';
 import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
 import { SocketDetailsMod } from 'app/item-popup/SocketDetails';
 import React from 'react';
-import ClosableContainer from '../ClosableContainer';
+import ClosableContainer from './ClosableContainer';
 import styles from './LockedModIcon.m.scss';
 
 interface Props {
