+++ src/app/loadout/mod-picker/ModPickerFooter.m.scss.d.ts
