+++ src/app/loadout/mod-picker/ModPickerSection.m.scss
