+++ src/app/loadout/mod-picker/SelectableMod.m.scss
